+++ embassy-stm32/src/timer/one_pulse.rs
@@ -15,6 +15,7 @@ use crate::gpio::{AfType, AnyPin, Pull};
 use crate::interrupt::typelevel::{Binding, Interrupt};
 use crate::pac::timer::vals::Etp;
 use crate::time::Hertz;
+use crate::timer::low_level::mode;
 use crate::Peri;
 
 /// External input marker type.
@@ -123,7 +124,7 @@ impl<'d, T: GeneralInstance4Channel, C: TriggerSource> TriggerPin<'d, T, C> {
 ///
 /// Generates a pulse after a trigger and some configurable delay.
 pub struct OnePulse<'d, T: GeneralInstance4Channel> {
-    inner: Timer<'d, T>,
+    inner: Timer<'d, T, mode::NoIrq>,
 }
 
 impl<'d, T: GeneralInstance4Channel> OnePulse<'d, T> {
@@ -354,7 +355,7 @@ pub struct OnePulseChannels<'d, T: GeneralInstance4Channel> {
 /// It is not possible to change the pulse end tick because the end tick
 /// configuration is shared with all four channels.
 pub struct OnePulseChannel<'d, T: GeneralInstance4Channel> {
-    inner: ManuallyDrop<Timer<'d, T>>,
+    inner: ManuallyDrop<Timer<'d, T, mode::NoIrq>>,
     channel: Channel,
 }